Link math tests using `sin()` to `libm`
While this linking seems to be done indirectly on most platforms, it
fails to work correctly on Sparc and S390x where an inappropriate
version of `sin()` is used. This should additional occurrences of bug #68.

Tests are passing on amd64, sparc and s390x Debian Linux platforms. Both
Python 2.x and Python 3.x.